README.md: add a README file with a quick buildroot intro
Although buildroot is useful, its documentation is sparse, sometimes
making it difficult to find specific answers. For example, the
information on replacing source directories could only be found if
someone knew to search for OVERRIDE_SRCDIR. Even then, the exact form
of the make targets was not well explained. Is it uppercase? Is it a
hyphen, or an underscore?

Document the parts of buildroot where I had the hardest time. Put
emphasis on _OVERRIDE_SRCDIR development, as that is the most likely
use case for this repo.

Signed-off-by: Alexandru Gagniuc <mr.nuke.me@gmail.com>